Initial support for linear framebuffer modes
It's at a point where it doesn't break anything, but doesn't really seem
to improve performance much either. At some point I will try to make the
driver treat the screen as a "huge" memory bitmap when using a linear
mode, which will hopefully allow more efficient codepaths to be used
(particularly for screen-to-screen blitting!).