Adding new binary SETUP.EXE for installing VBESVGA driver and performing initial configuration
SETUP.ASM references a VBESVGA.TXT that needs to be written at some point; it would basically
contain the same text as Configuration Parameters through the first half of Having Trouble
Finding Modes from README.md and the license info (whatever it is).

Note that the video mode picker is still the same guess as VIDMODES.COM and may need to be
tuned in the future (something about depth vs significant depth).