Fix even more clearly broken Vertical Retrace simulation in V_Stat0!
This code applies flags from the status 1 register to the status 0
register, which is obviously incorrect. The thing is, the code probably
never ran, because it's only set to run when the virtualized status 0
register is set to a non-zero value, which never happens anywhere else
in the driver anyway. But with the new VRTC code, that will change...