Improve framerate regulation for windowed DOS games
I have only tested this with Rayman, so other games may still be
completely broken. Ideally I would use paired timers to turn VRTC on AND
off, calculated based on the ratio of (VRTEnd - VRTStart) & 0Fh to
VTotal. But in practice that'd end up giving a period of less than a
millisecond for the "while VRTC on" timer, which is too short, so let's
leave it like this until someone complains...